Extract mod features
